hmr/requirements.txt
scipy
numpy
opendr==0.77
matplotlib
scikit-image
deepdish>=0.3
opencv-python
absl-py
ipdb
tensorflow==1.3.0
torch==1.0.0
